Moving files to a more logical location.
