+++ src/MarqueeSelection.ts
@@ -1,0 +1,3 @@
+export * from './components/MarqueeSelection/MarqueeSelection';
+export * from './components/MarqueeSelection/MarqueeSelection.Props';
+export * from './utilities/selection/index';
