+++ src/components/MarqueeSelection/MarqueeSelection.scss
