+++ src/utilities/AutoScroll/AutoScroll.ts
@@ -5,6 +5,12 @@ const SCROLL_ITERATION_DELAY = 30;
 const SCROLL_GUTTER_HEIGHT = 100;
 const MAX_SCROLL_VELOCITY = 20;
 
+/**
+ * AutoScroll simply hooks up mouse events given a parent element, and scrolls the container
+ * up/down depending on how close the mouse is to the top/bottom of the container.
+ *
+ * Once you don't want autoscroll any more, just dispose the helper and it will unhook events.
+ */
 export class AutoScroll {
   private _events: EventGroup;
   private _scrollableParent: HTMLElement;
