+++ src/utilities/selection/index.ts
@@ -2,4 +2,3 @@ export * from './interfaces';
 export * from './Selection';
 export * from './SelectionLayout';
 export * from './SelectionZone';
-export * from './MarqueeSelection';
